% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_binary_class.R
\name{create_binary_class}
\alias{create_binary_class}
\title{Returns a binary class}
\usage{
create_binary_class(point_info, var_name, use_id)
}
\arguments{
\item{point_info}{dataframe which has at column name that matches the var_name}

\item{var_name}{a string, that gives the variable name for matching}

\item{use_id}{the criteria for matching that generates the binary class}
}
\value{
Returns point_info with a new column called binary_class
}
\description{
Takes a data frame and finds the columns matching the variable name provided.
For the entries in that column, it creates a new variable, binary_class,
where the binary_class is given by var_name == use_id.
Function is useful for preprocessing data for fitting max-stable models.
}
\examples{
df = data.frame(x= 1:10, y = 1:10, z = rep(1:2, each = 5))
df_update = create_binary_class(df, "z", 2)

}
