% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_smith_model1.R
\name{fit_smith_model1}
\alias{fit_smith_model1}
\title{Fit smith max-stable model to bootstrap samples}
\usage{
fit_smith_model1(data, knn_info, medoid_indexes, use_id, grid, min_class_prob,
  min_stns = 10, min_common_obs, min_pairs = choose(10, 2),
  ratio_threshold = 0.1)
}
\arguments{
\item{data}{a data frame with columns labelled by the station id, and
rows corresponding to block maximum observation at that station}

\item{knn_info}{data frame with columns x, y, cluster_id, knn_id and prob.
The set of cluster_id and knn_id must be the same. The column prob is
the probability of assigning that station to under the classification.}

\item{medoid_indexes}{index of the medoid in the knn_info data frame.
should update this so it is a data frame for reference not an index.}

\item{use_id}{the cluster_id/knn_id of the region we are considering}

\item{grid}{a data frame that describes a grid, columns of x, y and knn_id.
Must cover the region of interest}

\item{min_class_prob}{for the station to be considered in the fitting it must
have been assigned under the classification with this minimum probability}

\item{min_common_obs}{minimum number of common observation for get_pair_weights()}

\item{min_pairs}{minimum number of pairs recommonded for fitting}

\item{ratio_threshold}{ratio of the elliptical axes must exceed this,
or convergence is supsected to have failed (default is 0.1, for 1:10)}

\item{num_stns}{minimum number of stations needed for fitting of the smith model (default = 10)}
}
\value{
Returns the a data frame with columns block, which is the year, and p_value.
}
\description{
We have classified a gridded region based on a clustering outpput. The classification
step was necessary to identify region boundaries. This function fits a max-stable model
to obtain the dependence structure for the region. We sample from suitable stations to
obtain multiple realisations of the dependence structure.
}
\details{
Of note, under the classification it is possible to have disjoint regions
classified similarly. For sampling, we only consider stations that are classified
similiarly and connected to the medoid identified in clustering. We also only
consider stations that were assigned to that region with a minimum probability.
After the initial fitting, we consider two basic diagnostics to check if the
optimisation converged. We compare the ratio of the two axes of the ellipses.
If the ratio is too small, we initial the start values and repeat the fitting.
We also repeat the fitting if the fitted parameter is 'too far'
from what we consider 'normal'.
}
