% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_grid.R
\name{generate_grid}
\alias{generate_grid}
\title{Generate a grid}
\usage{
generate_grid(coords, grid_space = 1)
}
\arguments{
\item{coords}{a data frame of coordinates with columns labelled x and y}

\item{grid_space}{the spacing for the grid, equal in both x and y directions
(defaults to 1)}
}
\value{
returns a data frame of gridded coordinates with columns labelled x and y
}
\description{
Function generates a grid to cover a set of given coordinates
(additional funcitonality to restrict the grid to inland areas of Australia)
}
\examples{
set.seed(1)
n.site <- 30
coords <-data.frame(x = runif(n.site, 0, 10), y = runif(n.site, 0, 10))
grid_output = generate_grid(coords = coords)
plot(grid_output)
points(coords, col = "red", pch = 20)

}
