% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classification_grid.R
\name{get_classification_grid}
\alias{get_classification_grid}
\title{Generate a grid}
\usage{
get_classification_grid(coords, grid_space = NULL, min_dist = NULL,
  restrict_aus = TRUE)
}
\arguments{
\item{coords}{a data frame of coordinates with columns labelled x and y}

\item{grid_space}{the spacing for the grid, equal in both x and y directions
(defaults to 1)}

\item{min_dist}{we restrict the grid to only cover the provided coordinates
any grid point greater than the min_dist away is removed (default is 1.5*grid_space)}

\item{restric_aus}{check if the coords fall within australia to
reduce computational time (default = TRUE)}
}
\value{
returns a data frame of gridded coordinates with columns labelled x and y
}
\description{
Function generates a grid to cover a set of given coordiantes
}
\examples{
n.site <- 30
coords <-data.frame(x = runif(n.site, 0, 10),
y = runif(n.site, 0, 10))
grid_output = get_classification_grid(coords, restrict_aus = FALSE)
plot(grid_output)

}
