% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ellipse_from_smith_model_fits.R
\name{get_ellipse_from_smith_model_list}
\alias{get_ellipse_from_smith_model_list}
\title{Converts GEV distrubted data to unit Frechet data
All functions are wrapped to handle any errors}
\usage{
get_ellipse_from_smith_model_list(model_list, medoid = data.frame(x = 0, y =
  0))
}
\arguments{
\item{model_list}{list of smith max-stable model fits}

\item{medoid}{a data frame with the x and y coordinates of the centre
of the ellipses, defualt is a data.frame(x = 0, y = 0)}
}
\value{
returns a data frame with columns sim_index, x, and y
the sim_index references each the ellipse associated with each max-stable model
}
\description{
Converts GEV distrubted data to unit Frechet data
All functions are wrapped to handle any errors
}
\examples{
n.site <- 30
locations <- matrix(runif(2*n.site, 0, 10), ncol = 2)
colnames(locations) <- c("lon", "lat")
data1 <- rmaxstab(100, locations, cov.mod = "gauss", cov11 = 1, cov12 = 0, cov22 = 1)
fit1 <- fitmaxstab(data1, locations, "gauss", loc.form <- loc ~ 1, scale.form <- scale ~ 1,
shape.form <- shape ~ 1,)
data2 <- rmaxstab(100, locations, cov.mod = "gauss", cov11 = 0.5, cov12 = 0, cov22 = 0.5)
fit2 <- fitmaxstab(data2, locations, "gauss", loc.form <- loc ~ 1, scale.form <- scale ~ 1,
shape.form <- shape ~ 1,)
model_list = list(fit1, fit2)
ellipse_list = get_ellipse_from_smith_model_list(model_list, medoid = data.frame(x = 1, y = 1))
ellipse_df = ellipse_list \%>\% as.data.frame()
ggplot(ellipse_df, aes(x = x, y=y, group = sim_index)) + geom_path()

}
