% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classification_grid.R
\name{get_grid_for_classification}
\alias{get_grid_for_classification}
\title{Generate a grid}
\usage{
get_grid_for_classification(coords, grid_space = 1, min_dist,
  poly_df = NULL)
}
\arguments{
\item{coords}{a data frame of coordinates with columns labelled x and y}

\item{grid_space}{the spacing for the grid, equal in both x and y directions
(defaults to 1)}

\item{min_dist}{we restrict the grid to only cover the provided coordinates
any grid point greater than the min_dist away is removed (default is 1.5*grid_space)}

\item{poly_df}{check if the coords fall within a polygon,
input is a data frame with columns x and y (default =  NULL)}
}
\value{
returns a data frame of gridded coordinates with columns labelled x and y
}
\description{
Function generates a grid to cover a set of given coordiantes
(additional funcitonality to restrict the grid to inland areas of Australia)
}
\examples{
set.seed(1)
n.site <- 30
coords <-data.frame(x = runif(n.site, 0, 10), y = runif(n.site, 0, 10))
grid_output = get_grid_for_classification(coords = coords)
plot(grid_output)
points(coords, col = "red", pch = 20)

poly_df <- data.frame(x = c(1,5,6,2,1), y = c(1,2,8,7,1))
grid_restricted = get_grid_for_classification(coords = coords, poly_df = poly_df)
plot(grid_output, pch = 3)
points(grid_restricted, col = "blue", pch = 3)
points(coords, col = "red", pch = 20)

}
