% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_medoid.R
\name{get_medoid}
\alias{get_medoid}
\title{Get a representative object for a set of points}
\usage{
get_medoid(coords, DD)
}
\arguments{
\item{coords}{data frame with columns x, y}

\item{DD}{a distance object related to the coordinates}
}
\value{
Returns a representative coordinates
}
\description{
This function find the point that minimises the distances to all the other
coordinates. This point is returned as a representative object
}
\examples{


coords = expand.grid(x = 1:9, y = 1:9)
DD = dist(coords, upper = TRUE, diag = FALSE)
medoid = get_medoid(coords, DD)
plot(coords)
points(medoid, col = "red", pch = 3)

set.seed(1)
coords = data.frame(x = runif(100), y = runif(100))
DD = dist(coords, upper = TRUE, diag = FALSE)
medoid = get_medoid(coords, DD)
plot(coords)
points(medoid, col = "red", pch = 3)

}
