% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples.R
\name{get_samples}
\alias{get_samples}
\title{Returns a set of samples from a vector}
\usage{
get_samples(n, sample_type, num_samples = 1, samp_size = NULL,
  num_partitions = NULL, percentage = NULL)
}
\arguments{
\item{n}{size of the vector we are sampling from}

\item{sample_type}{string of "random", "partition" or "percentage"}

\item{num_samples}{for random - number of sub-samples, default is 1}

\item{samp_size}{for random - size of the sub-sample to take}

\item{percentage}{for percentage - percentage of n that should be used as the samp_size.
In this instance, samp_size is calculated internally so does not need to be provided.}

\item{num_paritions}{for partition - number of ways to partition n and generate
sub-samples. If the num_partions does not divide n equally, the floor is taken.}
}
\value{
Returns a matrix where the columns give the sub-samples
}
\description{
For a vector of size n, sub-samples can be taken randomly, or n can be
partitioned into sub-samples.
}
\examples{
get_samples(n = 10, sample_type = "random", samp_size = 5)
get_samples(n = 10, sample_type = "random", num_samples = 2, samp_size = 5)
get_samples(n = 20, sample_type = "partition", num_partitions = 2)
get_samples(n = 10, sample_type = "percentage", num_samples = 2, percentage = 70)

}
