% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start_list.R
\name{get_start_list}
\alias{get_start_list}
\title{Get start list for fitmaxstab()}
\usage{
get_start_list(model_list, frech_bool = TRUE)
}
\arguments{
\item{model_list}{a list containing max-stable models}

\item{frech_bool}{(default = TRUE) are the marginals standard frechet}
}
\value{
returns a list of starting values for fitmaxstab()
}
\description{
Takes a list of models, and finds a set of starting values
}
\examples{

}
