% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev2frech_with_tryCatch.R
\name{gev2frech_with_tryCatch}
\alias{gev2frech_with_tryCatch}
\title{Converts GEV distrubted data to unit Frechet data
All functions are wrapped to handle any errors}
\usage{
gev2frech_with_tryCatch(x, gev_par = NULL)
}
\arguments{
\item{x}{a vector of block maximum observations}
}
\value{
returns x transformed to unit frechet data
}
\description{
Converts GEV distrubted data to unit Frechet data
All functions are wrapped to handle any errors
}
\examples{
x = revd(100, 0, 1, 0)
x_frech = gev2frech_with_tryCatch(x)
gev_par = fit_fevd_with_tryCatch(x)
x_frech1 = gev2frech_with_tryCatch(x, gev_par)
}
