% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cluster_connected.R
\name{check_clusters_connected}
\alias{check_clusters_connected}
\title{Check the classfied regions are connected}
\usage{
check_clusters_connected(coords, medoid, grid)
}
\arguments{
\item{coords}{data frame with columns x, y and class_id
class_id takes boolean values of 0 and 1}

\item{medoid}{index of the medoid in the coords data frame}

\item{grid}{a data frame that describes a grid, columsn of x, y and class_id}

\item{grid_space}{spacing of the grid, equal in x and y}
}
\value{
Returns the a data frame with columns block, which is the year, and p_value.
}
\description{
Under the classification, regions can be disconnected. 
For fitting max-stable models, we do not want to group stations in 
disjoint regions. This function returns as boolean if the station is density 
connected to it's medoid
}
\details{
For the two types of classification, we use 0 to denote the classfication 
we are interested in and 1 otherwise.
}
\examples{

coords = data.frame(x = runif(100)*10, y = runif(100)*10)
class_id = rep(1, nrow(coords))
coords = coords \%>\%
  mutate(class_id = class_id) \%>\%
  mutate(class_id = ifelse(x <= 3 & y <= 3, 0, class_id)) \%>\%
  mutate(class_id = ifelse(x >= 8 & y >= 2, 0, class_id))

medoid_coords = data.frame(x = 0.5, y = 0.5, class_id = 0)
coords = rbind(coords, medoid_coords)
medoid = nrow(coords)

coord_plot <-ggplot() +
  geom_point(data = coords, aes(x = x, y= y, 
                                col = as.factor(class_id))) +
  geom_point(data = medoid_coords, aes(x=x, y =y))

coord_plot

grid_space = 1
x_range = range(coords[,1])
y_range = range(coords[,2])
grid = expand.grid(x = seq(x_range[1], x_range[2], by = grid_space),
                   y = seq(y_range[1], y_range[2], by = grid_space))

training_set = coords \%>\% select(x,y) \%>\% as.matrix()
labels = coords \%>\% select(class_id) \%>\% unlist()
knn_classification = knn(train = training_set, 
                         test = grid, 
                         cl = labels, 
                         k = 1)

grid = mutate(grid, class_id = knn_classification)

grid_plot <-ggplot() +
  geom_point(data = coords, aes(x = x, y= y,
                                col = as.factor(class_id))) +
  geom_point(data = grid, aes(x = x, y = y,
                              col = as.factor(class_id)),
             shape = 3) +
  geom_point(data = medoid_coords, aes(x=x, y =y))

grid_plot

connected = check_clusters_connected(coords, medoid, grid)

connected_plot <- ggplot(data = coords, aes(x = x, y = y)) +
  geom_point(aes(col = as.factor(connected)))

connected_plot 
}
