% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classification_grid.R
\name{get_grid_for_classification}
\alias{get_grid_for_classification}
\title{Generate a grid}
\usage{
get_grid_for_classification(coords, grid_space = 1, min_dist,
  restrict_aus = TRUE)
}
\arguments{
\item{coords}{a data frame of coordinates with columns labelled x and y}

\item{grid_space}{the spacing for the grid, equal in both x and y directions
(defaults to 1)}

\item{min_dist}{we restrict the grid to only cover the provided coordinates
any grid point greater than the min_dist away is removed (default is 1.5*grid_space)}

\item{restrict_aus}{check if the coords fall within australia to
reduce computational time (default = TRUE)}
}
\value{
returns a data frame of gridded coordinates with columns labelled x and y
}
\description{
Function generates a grid to cover a set of given coordiantes
(additional funcitonality to restrict the grid to inland areas of Australia)
}
\examples{
set.seed(1)
n.site <- 30
coords <-data.frame(x = runif(n.site, 0, 10), y = runif(n.site, 0, 10))
grid_output = get_grid_for_classification(coords = coords, restrict_aus = FALSE)
plot(grid_output)
points(coords, col = "red", pch = 20)

n.site <- 100
coords <-data.frame(x = runif(n.site, 145, 155), y = runif(n.site, -25, -15))
grid_output = get_grid_for_classification(coords = coords, grid_space = 0.5)
oz(xlim = c(145,155), ylim = c(-25,-15))
points(grid_output)
points(coords, col = "red", pch = 20)
}
