% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outer_wrapper_fitmaxstab.R
\name{outer_wrapper_fitmaxstab}
\alias{outer_wrapper_fitmaxstab}
\title{Fit smith max-stable model}
\usage{
outer_wrapper_fitmaxstab(fit_info, obs_data, convert = FALSE,
  frech_bool = TRUE, cov_mod, min_common_obs = 10, min_pairs = 10,
  fit_subsample = FALSE, sample_type = NULL, ...)
}
\arguments{
\item{fit_info}{a data frame which must have columns id, x, y}

\item{obs_data}{a data frame with columns labelled by the station id, and
rows corresponding to block maximum observation at that station,
default assumption is this data is standard frechet}

\item{convert}{boolean to convert the data to standard frechet from GEV
distributed data (default = FALSE)}

\item{frech_bool}{(default = TRUE) is the data frechet distributed, otherwise we
assume loc, scale and shape parameters constant}

\item{cov_mod}{A character string corresponding to the covariance model in the
max-stable representation. See SpatialExtremes::fitmaxstab() documentation.}

\item{min_common_obs}{minimum number of common pairwise observations,
weight is set to zero in fitting otherwise (default is 10)}

\item{min_pairs}{set a minimum number of pairs needed for fitting (default is 10)}

\item{fit_subsample}{repeat fitting for sub samples of stations (default is FALSE)}

\item{sample_type}{type of samples to take, see get_samples() for documentation,
and addition parameters that need to be specified (Default is NULL, for no subsamples)}
}
\value{
Returns a list of fitted max-stable models
}
\description{
Fit smith max-stable model
}
\examples{
n.site <- 60
locations <- matrix(runif(2*n.site, 0, 10), ncol = 2)
id <- paste("ID", as.character(1:n.site), sep = "")
fit_info <- data.frame(x = locations[,1],
                       y = locations[,2],
                       id,
                       stringsAsFactors = FALSE)
sim_data <- SpatialExtremes::rmaxstab(100, locations, cov.mod = "gauss",
                  cov11 = 1, cov12 = 0, cov22 = 1)
sim_data <- data.frame(sim_data)
names(sim_data) <- id
fitM = outer_wrapper_fitmaxstab(fit_info = fit_info, obs_data = sim_data,
                cov_mod = "gauss")

}
